feat(but): replace old squash command with squash2
