--- crates/but/src/command/legacy/rub/squash.rs
@@ -1,581 +1,0 @@
-use anyhow::{Context as _, bail};
-use bstr::BString;
-use but_core::{DryRun, ref_metadata::StackId, sync::RepoExclusive};
-use but_ctx::Context;
-use but_workspace::commit::squash_commits::MessageCombinationStrategy;
-use gitbutler_oplog::{
-    OplogExt,
-    entry::{OperationKind, SnapshotDetails},
-};
-use gix::ObjectId;
-
-use crate::theme::{self, Paint};
-
-use super::undo::stack_id_by_commit_id;
-use crate::{CliId, IdMap, command::legacy::ai, id::CommitId, utils::OutputChannel};
-
-/// Handler for `but squash` command with support for:
-/// 1. Multiple commits: `but squash <commit1> <commit2> <commit3>` - squashes 1 and 2 into 3
-/// 2. Commit range: `but squash <commit1>..<commit4>` - squashes all in range into last
-/// 3. Branch name: `but squash <branch>` - squashes all commits in branch into bottom-most
-pub(crate) fn handle(
-    ctx: &mut Context,
-    out: &mut OutputChannel,
-    commits: &[String],
-    drop_message: bool,
-    custom_message: Option<&str>,
-    ai: Option<Option<String>>,
-) -> anyhow::Result<()> {
-    if commits.is_empty() {
-        bail!("At least one commit or branch name must be provided");
-    }
-
-    let mut guard = ctx.exclusive_worktree_access();
-    let id_map = IdMap::new_from_context(ctx, None, guard.read_permission())?;
-
-    // If there's only one argument, it could be a branch name or a range
-    if commits.len() == 1 {
-        let entity_str = &commits[0];
-
-        // First try to resolve as a single entity (branch name or commit)
-        let matches = id_map.parse_using_context(entity_str, ctx)?;
-
-        // If we get exactly one match, handle it
-        if matches.len() == 1 {
-            let entity = &matches[0];
-
-            // Check if it's a branch - if so, squash all commits in the branch
-            if let CliId::Branch(branch) = entity {
-                return squash_branch_commits(
-                    ctx,
-                    out,
-                    &branch.name,
-                    branch.stack_id,
-                    drop_message,
-                    custom_message,
-                    ai,
-                    &id_map,
-                    guard.write_permission(),
-                );
-            }
-
-            // If it's a single commit, error - need at least 2 commits
-            if let CliId::Commit { .. } = entity {
-                bail!(
-                    "Need at least 2 commits to squash. To squash all commits in a branch, use the branch name."
-                );
-            }
-
-            bail!("'{entity_str}' must be a branch name or commit identifier");
-        }
-
-        // If we get multiple matches, it's ambiguous
-        if matches.len() > 1 {
-            bail!("'{entity_str}' is ambiguous - matches multiple entities");
-        }
-
-        // No exact match found - try parsing as a range or list if it contains special characters
-        if entity_str.contains("..") {
-            let sources = parse_commit_range(ctx, &id_map, entity_str)?;
-            if sources.len() < 2 {
-                bail!("Need at least 2 commits to squash");
-            }
-            return handle_multi_commit_squash(
-                ctx,
-                &id_map,
-                out,
-                sources,
-                drop_message,
-                custom_message,
-                ai,
-                guard.write_permission(),
-            );
-        }
-
-        if entity_str.contains(',') {
-            let sources = parse_commit_list(ctx, &id_map, entity_str)?;
-            if sources.len() < 2 {
-                bail!("Need at least 2 commits to squash");
-            }
-            return handle_multi_commit_squash(
-                ctx,
-                &id_map,
-                out,
-                sources,
-                drop_message,
-                custom_message,
-                ai,
-                guard.write_permission(),
-            );
-        }
-
-        // If it contains a single dash (but not ..), it might be a branch name with a dash
-        // This case is already handled above by exact match, so we get here only if no match found
-
-        // Nothing worked
-        bail!("No matching branch or commit found for '{entity_str}'");
-    }
-
-    // Multiple separate commit arguments - resolve each one
-    let mut sources = Vec::new();
-    for commit_str in commits {
-        let matches = id_map.parse_using_context(commit_str, ctx)?;
-        if matches.is_empty() {
-            bail!("No matching commit found for '{commit_str}'");
-        }
-        if matches.len() > 1 {
-            bail!("'{commit_str}' is ambiguous - matches multiple entities");
-        }
-        sources.push(matches[0].clone());
-    }
-
-    handle_multi_commit_squash(
-        ctx,
-        &id_map,
-        out,
-        sources,
-        drop_message,
-        custom_message,
-        ai,
-        guard.write_permission(),
-    )
-}
-
-/// Helper function to handle squashing multiple commits
-#[expect(clippy::too_many_arguments)]
-fn handle_multi_commit_squash(
-    ctx: &mut Context,
-    id_map: &IdMap,
-    out: &mut OutputChannel,
-    sources: Vec<CliId>,
-    drop_message: bool,
-    custom_message: Option<&str>,
-    ai: Option<Option<String>>,
-    perm: &mut RepoExclusive,
-) -> anyhow::Result<()> {
-    if sources.len() < 2 {
-        bail!("Need at least 2 commits to squash");
-    }
-
-    // Extract commit OIDs and validate all are commits
-    let mut commit_oids = Vec::new();
-    for source in &sources {
-        match source {
-            CliId::Commit(CommitId { commit_id, .. }) => {
-                commit_oids.push(*commit_id);
-            }
-            other => {
-                bail!(
-                    "Cannot squash {} - it is {}. All arguments must be commits.",
-                    theme::get().cli_id.paint(other.to_short_string()),
-                    theme::get().attention.paint(other.kind_for_humans())
-                );
-            }
-        }
-    }
-
-    // The last commit is the target
-    let target_oid = commit_oids.pop().expect("We validated sources.len() >= 2");
-
-    // Delegate to the shared squashing logic
-    squash_commits_internal(
-        ctx,
-        id_map,
-        commit_oids,
-        target_oid,
-        drop_message,
-        custom_message,
-        ai,
-        perm,
-        out,
-    )
-}
-
-/// Internal shared logic for squashing commits
-#[expect(clippy::too_many_arguments)]
-fn squash_commits_internal(
-    ctx: &mut Context,
-    id_map: &IdMap,
-    source_oids: Vec<ObjectId>,
-    target_oid: ObjectId,
-    drop_message: bool,
-    custom_message: Option<&str>,
-    ai: Option<Option<String>>,
-    perm: &mut RepoExclusive,
-    out: &mut OutputChannel,
-) -> anyhow::Result<()> {
-    // Collect commit messages if we need them for AI generation
-    let (source_messages, destination_message) = if ai.is_some() {
-        let repo = ctx.repo.get()?;
-
-        // Get source commit messages
-        let mut source_msgs = Vec::new();
-        for source_oid in &source_oids {
-            let commit = repo.find_commit(*source_oid)?;
-            let message_ref = commit.message()?;
-            let full_message = if let Some(body) = message_ref.body {
-                format!("{}\n\n{}", message_ref.title, body)
-            } else {
-                message_ref.title.to_string()
-            };
-            source_msgs.push(full_message);
-        }
-
-        // Get destination commit message
-        let target_commit = repo.find_commit(target_oid)?;
-        let message_ref = target_commit.message()?;
-        let dest_message = if let Some(body) = message_ref.body {
-            format!("{}\n\n{}", message_ref.title, body)
-        } else {
-            message_ref.title.to_string()
-        };
-
-        (Some(source_msgs), Some(dest_message))
-    } else {
-        (None, None)
-    };
-
-    // If drop_message is set, get the target commit's message BEFORE squashing
-    let target_message = if drop_message {
-        let repo = ctx.repo.get()?;
-        let target_commit = repo.find_commit(target_oid)?;
-        let message_ref = target_commit.message()?;
-        let full_message = if let Some(body) = message_ref.body {
-            format!("{}\n\n{}", message_ref.title, body)
-        } else {
-            message_ref.title.to_string()
-        };
-        Some(full_message)
-    } else {
-        None
-    };
-
-    // Keep the complete multi-rewrite sequence atomic: if any step fails,
-    // restore the pre-squash snapshot so we don't leave partial rewrites.
-    let snapshot = ctx.create_snapshot(SnapshotDetails::new(OperationKind::SquashCommit), perm)?;
-    let source_oids_count = source_oids.len();
-    let only_source_commit = source_oids.first().copied();
-    let target_change_id = id_map
-        .change_id_ref(target_oid)
-        .map(|change_id| change_id.change_id.clone());
-    let source_oids_to_squash = source_oids;
-
-    let squash_result: anyhow::Result<ObjectId> = (|| {
-        let squash_result = but_api::commit::squash::commit_squash_only_with_perm(
-            ctx,
-            source_oids_to_squash,
-            target_oid,
-            MessageCombinationStrategy::KeepBoth,
-            DryRun::No,
-            perm,
-        )?;
-        let new_commit_oid = squash_result.new_commit;
-
-        // Determine the final message and apply if needed.
-        let final_commit_oid = if let Some(user_summary) = ai {
-            // Use AI to generate the commit message.
-            let ai_message = ai::generate_commit_message_from_multiple_messages(
-                out,
-                source_messages.unwrap_or_default(),
-                destination_message.unwrap_or_default(),
-                user_summary,
-            )?;
-            but_api::commit::reword::commit_reword_only_with_perm(
-                ctx,
-                new_commit_oid,
-                BString::from(ai_message),
-                DryRun::No,
-                perm,
-            )?
-            .new_commit
-        } else if let Some(msg) = custom_message {
-            but_api::commit::reword::commit_reword_only_with_perm(
-                ctx,
-                new_commit_oid,
-                BString::from(msg),
-                DryRun::No,
-                perm,
-            )?
-            .new_commit
-        } else if let Some(target_msg) = target_message {
-            but_api::commit::reword::commit_reword_only_with_perm(
-                ctx,
-                new_commit_oid,
-                BString::from(target_msg),
-                DryRun::No,
-                perm,
-            )?
-            .new_commit
-        } else {
-            new_commit_oid
-        };
-
-        Ok(final_commit_oid)
-    })();
-
-    let final_commit_oid = match squash_result {
-        Ok(final_commit_oid) => final_commit_oid,
-        Err(err) => {
-            ctx.restore_snapshot(
-                snapshot,
-                gitbutler_oplog::RestoreKind::ExplicitRestoreFromSnapshot,
-                perm,
-            )
-            .with_context(|| {
-                format!("Failed to restore snapshot {snapshot} after squash failure")
-            })?;
-            return Err(err);
-        }
-    };
-
-    // Output message based on context
-    if let Some(out) = out.for_human() {
-        let new_commit = theme::Commit(final_commit_oid, target_change_id);
-        if source_oids_count == 1 {
-            // Single commit squash (for backwards compatibility with `but rub`)
-            let source = only_source_commit
-                .context("BUG: Source commits count is one, but first item is none")?;
-            writeln!(
-                out,
-                "Squashed {} → {}",
-                theme::Commit(
-                    source,
-                    id_map
-                        .change_id_ref(source)
-                        .map(|change_id| change_id.change_id.clone()),
-                ),
-                new_commit
-            )?
-        } else {
-            // Multiple commits squash
-            writeln!(out, "Squashed {source_oids_count} commits → {new_commit}")?
-        }
-    } else if let Some(out) = out.for_json() {
-        out.write_value(serde_json::json!({
-            "ok": true,
-            "new_commit_id": final_commit_oid.to_string(),
-            "squashed_count": source_oids_count,
-        }))?;
-    }
-    Ok(())
-}
-
-/// Helper function to squash all commits in a branch into the bottom-most commit
-#[expect(clippy::too_many_arguments)]
-fn squash_branch_commits(
-    ctx: &mut Context,
-    out: &mut OutputChannel,
-    branch_name: &str,
-    stack_id: Option<StackId>,
-    drop_message: bool,
-    custom_message: Option<&str>,
-    ai: Option<Option<String>>,
-    id_map: &IdMap,
-    perm: &mut RepoExclusive,
-) -> anyhow::Result<()> {
-    // Find the stack containing this branch
-    let stack_id = stack_id
-        .ok_or_else(|| anyhow::anyhow!("Branch '{branch_name}' is not associated with a stack"))?;
-
-    // Find all commits in this branch (segment)
-    let mut branch_commits: Vec<gix::ObjectId> = Vec::new();
-    for stack in id_map.stacks() {
-        if stack.id == Some(stack_id) {
-            for segment in &stack.segments {
-                if let Some(seg_branch_name) = segment.branch_name()
-                    && seg_branch_name == branch_name.as_bytes()
-                {
-                    // Collect all workspace commits in this segment
-                    for commit in &segment.workspace_commits {
-                        branch_commits.push(commit.commit_id());
-                    }
-                    break;
-                }
-            }
-            break;
-        }
-    }
-
-    if branch_commits.is_empty() {
-        bail!("No commits found in branch '{branch_name}'");
-    }
-
-    if branch_commits.len() < 2 {
-        bail!("Branch '{branch_name}' has only one commit, nothing to squash");
-    }
-
-    // The commits are in order from newest (top) to oldest (bottom)
-    // We want to squash all commits into the bottom-most (last in the list)
-    let target_oid = branch_commits.pop().expect("We validated len >= 2");
-    let source_oids = branch_commits; // These are already ObjectIds
-
-    // Delegate to the shared squashing logic
-    squash_commits_internal(
-        ctx,
-        id_map,
-        source_oids,
-        target_oid,
-        drop_message,
-        custom_message,
-        ai,
-        perm,
-        out,
-    )?;
-
-    // Add branch-specific output message
-    let t = theme::get();
-    if let Some(out) = out.for_human() {
-        writeln!(
-            out,
-            "Squashed all commits in branch '{}'",
-            t.local_branch.paint(branch_name)
-        )?
-    }
-    Ok(())
-}
-
-/// Parse a commit range like "c1..c3" and return all commits in the range
-fn parse_commit_range(
-    ctx: &mut Context,
-    id_map: &IdMap,
-    range_str: &str,
-) -> anyhow::Result<Vec<CliId>> {
-    let parts: Vec<&str> = range_str.split("..").collect();
-    if parts.len() != 2 {
-        bail!("Range format should be 'start..end', got '{range_str}'");
-    }
-
-    let start_str = parts[0];
-    let end_str = parts[1];
-
-    // Resolve start and end to commit IDs
-    let start_matches = id_map.parse_using_context(start_str, ctx)?;
-    let end_matches = id_map.parse_using_context(end_str, ctx)?;
-
-    if start_matches.len() != 1 {
-        bail!("Start of range '{start_str}' must match exactly one commit");
-    }
-    if end_matches.len() != 1 {
-        bail!("End of range '{end_str}' must match exactly one commit");
-    }
-
-    let start_id = &start_matches[0];
-    let end_id = &end_matches[0];
-
-    // Both must be commits
-    let (start_commit_oid, end_commit_oid) = match (start_id, end_id) {
-        (
-            CliId::Commit(CommitId {
-                commit_id: start_oid,
-                ..
-            }),
-            CliId::Commit(CommitId {
-                commit_id: end_oid, ..
-            }),
-        ) => (start_oid, end_oid),
-        _ => {
-            bail!("Range endpoints must be commits, not other types");
-        }
-    };
-
-    // Verify both commits are on the same stack FIRST
-    let start_stack = stack_id_by_commit_id(ctx, *start_commit_oid)?;
-    let end_stack = stack_id_by_commit_id(ctx, *end_commit_oid)?;
-    if start_stack != end_stack {
-        bail!(
-            "Range endpoints must be on the same stack. '{start_str}' and '{end_str}' are on different stacks."
-        );
-    }
-
-    // Get all commits in order from the SAME stack only
-    let mut all_commits_in_order: Vec<(gix::ObjectId, CliId)> = Vec::new();
-    for stack in id_map.stacks() {
-        // Only process the stack that contains our commits
-        if stack.id != Some(start_stack) {
-            continue;
-        }
-
-        for segment in &stack.segments {
-            for commit in &segment.workspace_commits {
-                let commit_oid = commit.commit_id();
-                // Find the CliId for this commit
-                if let Some(cli_id) = id_map
-                    .parse_using_context(&commit_oid.to_string(), ctx)
-                    .ok()
-                    .and_then(|matches| matches.first().cloned())
-                {
-                    all_commits_in_order.push((commit_oid, cli_id));
-                }
-            }
-        }
-        break; // Found our stack, no need to continue
-    }
-
-    // Find the positions of start and end commits
-    let start_pos = all_commits_in_order
-        .iter()
-        .position(|(oid, _)| oid == start_commit_oid);
-    let end_pos = all_commits_in_order
-        .iter()
-        .position(|(oid, _)| oid == end_commit_oid);
-
-    match (start_pos, end_pos) {
-        (Some(start_idx), Some(end_idx)) => {
-            let range = if start_idx <= end_idx {
-                &all_commits_in_order[start_idx..=end_idx]
-            } else {
-                &all_commits_in_order[end_idx..=start_idx]
-            };
-
-            Ok(range.iter().map(|(_, cli_id)| cli_id.clone()).collect())
-        }
-        _ => {
-            bail!(
-                "Could not find range from '{start_str}' to '{end_str}'. Make sure both commits exist in the stack."
-            );
-        }
-    }
-}
-
-/// Parse a comma-separated list of commits like "c1,c2,c3"
-fn parse_commit_list(
-    ctx: &mut Context,
-    id_map: &IdMap,
-    list_str: &str,
-) -> anyhow::Result<Vec<CliId>> {
-    let parts: Vec<&str> = list_str.split(',').collect();
-    let mut result = Vec::new();
-
-    for part in parts {
-        let part = part.trim();
-        if part.is_empty() {
-            continue;
-        }
-
-        let matches = id_map.parse_using_context(part, ctx)?;
-        if matches.is_empty() {
-            bail!("Commit '{part}' not found");
-        }
-        if matches.len() > 1 {
-            bail!("'{part}' is ambiguous - matches multiple entities");
-        }
-
-        // Verify it's a commit
-        match &matches[0] {
-            CliId::Commit { .. } => result.push(matches[0].clone()),
-            other => {
-                bail!(
-                    "'{}' is {} but must be a commit",
-                    part,
-                    other.kind_for_humans()
-                );
-            }
-        }
-    }
-
-    if result.is_empty() {
-        bail!("Commit list is empty");
-    }
-
-    Ok(result)
-}
