--- crates/but/tests/but/command/squash2.rs
@@ -1,2454 +1,0 @@
-use snapbox::str;
-
-use crate::{
-    command::util::{
-        branch_commit_cli_id_for_file, branch_commit_cli_ids,
-        commit_file_with_worktree_changes_as_two_hunks, commit_two_files_as_two_hunks_each,
-        status_json_with_files as status_json,
-    },
-    utils::{CommandExt, Sandbox},
-};
-
-fn uncommitted_contains_file(status: &serde_json::Value, file_path: &str) -> bool {
-    status["uncommittedChanges"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .any(|change| change["filePath"].as_str().unwrap() == file_path)
-}
-
-fn uncommitted_cli_id_for_file(status: &serde_json::Value, file_path: &str) -> Option<String> {
-    status["uncommittedChanges"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .find_map(|change| {
-            (change["filePath"].as_str().unwrap() == file_path)
-                .then(|| change["cliId"].as_str().unwrap().to_string())
-        })
-}
-
-fn branch_commits_contain_file(
-    status: &serde_json::Value,
-    branch_name: &str,
-    file_path: &str,
-) -> bool {
-    status["stacks"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .flat_map(|stack| stack["branches"].as_array().unwrap().iter())
-        .filter(|branch| branch["name"].as_str().unwrap() == branch_name)
-        .flat_map(|branch| branch["commits"].as_array().unwrap().iter())
-        .flat_map(|commit| commit["changes"].as_array().unwrap().iter())
-        .any(|change| change["filePath"].as_str().unwrap() == file_path)
-}
-
-fn one_branch_three_commits() -> Sandbox {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("one", "content of one");
-    env.file("two", "content of two");
-    env.file("three", "content of three");
-
-    env.but("commit -m 'add one' one").assert().success();
-    env.but("commit -m 'add two' two").assert().success();
-    env.but("commit -m 'add three' three").assert().success();
-
-    env
-}
-
-fn two_branches() -> Sandbox {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("one", "content of one");
-    env.file("two", "content of two");
-    env.file("three", "content of three");
-    env.file("four", "content of four");
-
-    env.but("commit -b one -m 'add one' one").assert().success();
-    env.but("commit -b one -m 'add two' two").assert().success();
-
-    env.but("commit -b second -m 'add three' three")
-        .assert()
-        .success();
-    env.but("commit -b second -m 'add four' four")
-        .assert()
-        .success();
-
-    env
-}
-
-fn scenario_with_uncommitted_changes() -> Sandbox {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("one", "content of one");
-    env.file("two", "content of two");
-    env.file("three", "content of three");
-
-    env.but("commit --empty --no-message").assert().success();
-
-    env
-}
-
-#[test]
-fn squash_two_commits() {
-    let env = one_branch_three_commits();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three
-┊│     1#0:o A three
-┊●   1#1 add two
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#0 --target 1#1 --message 'squashed'")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed f55169f into f63361f to create 7251301
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 squashed
-┊│     1#0:o A three
-┊│     1#0:t A two
-┊●   1#1 add one
-┊│     1#1:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("undo").assert().success();
-
-    env.but("_squash2 1#0 --target 1#1 --message 'squashed' --format json")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-{
-  "new_commit": "725130139e9f0178e29afbe9eff6a988afbca3fa"
-}
-
-"#]]);
-
-    env.but("undo").assert().success();
-
-    env.but("_squash2 1#0 --target 1#1 --message 'squashed' --format shell")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-725130139e9f0178e29afbe9eff6a988afbca3fa
-
-"#]]);
-}
-
-#[test]
-fn squash_multiple_sources() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 1#1 --target 1#2 --message 'squashed'")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed f55169f, f63361f into ea345ba to create e355a10
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 squashed
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn use_target_message() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 --target 1#1 --use-target-message")
-        .assert()
-        .success();
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊● 1#0 author 2000-01-01 00:00:00 +0000 (sha 5ab5165)
-┊│     add two
-┊│     1#0:o A three
-┊│     1#0:t A two
-┊● 1#1 author 2000-01-01 00:00:00 +0000 (sha ea345ba)
-┊│     add one
-┊│     1#1:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn use_source_message() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 --target 1#1 --use-source-message")
-        .assert()
-        .success();
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊● 1#0 author 2000-01-01 00:00:00 +0000 (sha c441d34)
-┊│     add three
-┊│     1#0:o A three
-┊│     1#0:t A two
-┊● 1#1 author 2000-01-01 00:00:00 +0000 (sha ea345ba)
-┊│     add one
-┊│     1#1:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_whole_branch() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 a-branch-1 -m 'squashed a branch'")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![["
-Squashed branch 'a-branch-1' to create commit a694042
-
-"]]);
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊● 1 author 2000-01-01 00:00:00 +0000 (sha a694042)
-┊│     squashed a branch
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_whole_branch_into_commit_on_same_branch() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 a-branch-1 -t 1#1 --use-target-message")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed branch 'a-branch-1' to create commit 17b59a2
-
-"#]]);
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊● 1 author 2000-01-01 00:00:00 +0000 (sha 17b59a2)
-┊│     add two
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_whole_branch_into_commit_on_other_branch() {
-    let env = one_branch_three_commits();
-
-    env.but("commit -b target-branch -m 'new commit on new branch'")
-        .assert()
-        .success();
-
-    env.file("file", "new file");
-    env.but("commit file -b add-file-branch -m 'add file'")
-        .assert()
-        .success();
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ fi [add-file-branch]
-┊● 1#0 author 2000-01-01 00:00:00 +0000 (sha e528488)
-┊│     add file
-┊│     1#0:q A file
-├╯
-┊
-┊╭┄ ta [target-branch]
-┊● 1#1 author 2000-01-01 00:00:00 +0000 (sha d1d6a19) (no changes)
-┊│     new commit on new branch
-├╯
-┊
-┊╭┄ br [a-branch-1]
-┊● 1#2 author 2000-01-01 00:00:00 +0000 (sha f55169f)
-┊│     add three
-┊│     1#2:o A three
-┊● 1#3 author 2000-01-01 00:00:00 +0000 (sha f63361f)
-┊│     add two
-┊│     1#3:t A two
-┊● 1#4 author 2000-01-01 00:00:00 +0000 (sha ea345ba)
-┊│     add one
-┊│     1#4:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 a-branch-1 add-file-branch -t 1#1 --use-target-message")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed branches 'a-branch-1', 'add-file-branch' to create commit 44aa30a
-
-"#]]);
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ ta [target-branch]
-┊● 1 author 2000-01-01 00:00:00 +0000 (sha 44aa30a)
-┊│     new commit on new branch
-┊│     1:q A file
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_multiple_branches_into_commit_on_one_of_the_branch_sources() {
-    let env = one_branch_three_commits();
-
-    env.but("commit -b target-branch -m 'target commit'")
-        .assert()
-        .success();
-    env.but("commit -b target-branch -m 'random commit on target-branch'")
-        .assert()
-        .success();
-
-    env.file("file", "new file");
-    env.but("commit file -b add-file-branch -m 'add file'")
-        .assert()
-        .success();
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ fi [add-file-branch]
-┊● 1#0 author 2000-01-01 00:00:00 +0000 (sha e528488)
-┊│     add file
-┊│     1#0:q A file
-├╯
-┊
-┊╭┄ ta [target-branch]
-┊● 1#1 author 2000-01-01 00:00:00 +0000 (sha a489b93) (no changes)
-┊│     random commit on target-branch
-┊● 1#2 author 2000-01-01 00:00:00 +0000 (sha 561a8d8) (no changes)
-┊│     target commit
-├╯
-┊
-┊╭┄ br [a-branch-1]
-┊● 1#3 author 2000-01-01 00:00:00 +0000 (sha f55169f)
-┊│     add three
-┊│     1#3:o A three
-┊● 1#4 author 2000-01-01 00:00:00 +0000 (sha f63361f)
-┊│     add two
-┊│     1#4:t A two
-┊● 1#5 author 2000-01-01 00:00:00 +0000 (sha ea345ba)
-┊│     add one
-┊│     1#5:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 target-branch a-branch-1 add-file-branch -t 1#2 --use-target-message")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed branches 'target-branch', 'a-branch-1', 'add-file-branch' to create commit 0653794
-
-"#]]);
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ ta [target-branch]
-┊● 1 author 2000-01-01 00:00:00 +0000 (sha 0653794)
-┊│     target commit
-┊│     1:q A file
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_reword_with_editor() {
-    let env = one_branch_three_commits();
-
-    env.file(
-        ".git/editor.sh",
-        "printf 'message from editor\\n' > \"$1\"\n",
-    );
-    let editor_path = env.projects_root().join(".git/editor.sh");
-    let editor_command = format!("sh {}", editor_path.display());
-
-    env.but("_squash2 a-branch-1")
-        .env("GIT_EDITOR", editor_command)
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![["
-Squashed branch 'a-branch-1' to create commit 7b3d915
-
-"]]);
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊● 1 author 2000-01-01 00:00:00 +0000 (sha 7b3d915)
-┊│     message from editor
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_combine_messages_with_editor() {
-    let env = one_branch_three_commits();
-
-    env.file(".git/editor.sh", "true");
-    let editor_path = env.projects_root().join(".git/editor.sh");
-    let editor_command = format!("sh {}", editor_path.display());
-
-    env.but("_squash2 a-branch-1")
-        .env("GIT_EDITOR", editor_command)
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed branch 'a-branch-1' to create commit abb21d9
-
-"#]]);
-
-    env.but("status -fv")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊● 1 author 2000-01-01 00:00:00 +0000 (sha abb21d9)
-┊│     add one  add three  add two
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_nothing() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-error: the following required arguments were not provided:
-  <SOURCES>...
-
-Usage: but _squash2 <SOURCES>...
-
-For more information, try '--help'.
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_only_target() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 --target 1#0")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-error: the following required arguments were not provided:
-  <SOURCES>...
-
-Usage: but _squash2 --target <TARGET> <SOURCES>...
-
-For more information, try '--help'.
-
-"#]]);
-}
-
-#[test]
-fn cannot_mix_sources() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 a-branch-1 1#0 --target 1#2")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Cannot mix different types of sources
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_multiple_commits_without_target() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 1#2")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: When --target isn't used the source must be exactly one branch
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_multiple_branches_without_target() {
-    let env = one_branch_three_commits();
-
-    env.but("commit --no-message -b second-branch")
-        .assert()
-        .success();
-
-    env.but("_squash2 a-branch-1 second-branch")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: When --target isn't used the source must be exactly one branch
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_branch_with_just_one_commit() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("one", "content of one");
-    env.but("commit -m 'add one' one -b the-branch")
-        .assert()
-        .success();
-
-    env.but("_squash2 the-branch -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Need at least 2 commits to squash
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_commit_into_itself() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 -t 1#0")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Cannot squash a commit into itself
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_empty_branch_into_itself() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.but("branch new empty-branch").assert().success();
-
-    env.but("_squash2 empty-branch")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Cannot squash empty branch into itself
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_empty_branch_into_commit() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.but("commit -m 'target commit'").assert().success();
-
-    env.but("branch new empty-branch").assert().success();
-
-    env.but("_squash2 empty-branch -t 1")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Need at least 2 commits to squash
-
-"#]]);
-}
-
-#[test]
-fn aborts_on_conflicts() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file.txt", "file content");
-    env.but("commit -m 'add file'").assert().success();
-
-    env.file("file.txt", "changed file content");
-    env.but("commit -m 'change file'").assert().success();
-
-    env.remove_file("file.txt");
-    env.but("commit -m 'remove file'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 remove file
-┊│     1#0:u D file.txt
-┊●   1#1 change file
-┊│     1#1:u M file.txt
-┊●   1#2 add file
-┊│     1#2:u A file.txt
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#0 -t 1#2")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Cannot squash commits that would result in merge conflicts
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_into_commits_on_unapplied_branches() {
-    let env = two_branches();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ se [second]
-┊●   1#0 add four
-┊│     1#0:q A four
-┊●   1#1 add three
-┊│     1#1:o A three
-├╯
-┊
-┊╭┄ on [one]
-┊●   1#2 add two
-┊│     1#2:t A two
-┊●   1#3 add one
-┊│     1#3:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("unapply second").assert().success();
-
-    // Unapplied commits have no change ID in the workspace map, so use the commit ID intentionally.
-    env.but("_squash2 1#0 -t d15f721")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Could not find target: 'd15f721'
-
-Hint: --target must be an applied commit, branch, or zz. Run `but status` for applicable targets.
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_unapplied_branch() {
-    let env = two_branches();
-
-    env.but("unapply second").assert().success();
-
-    env.but("_squash2 second")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Could not find branch: 'second'
-
-Hint: Run `but status` for applicable targets.
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_branch_with_one_commit_into_that_one_commit() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 A -t tpm")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Need at least 2 commits to squash
-
-"#]]);
-}
-
-#[test]
-fn squash_with_duplicate_commit_sources() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 1#0 -t 1#1 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed f55169f into f63361f to create 5ab5165
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add two
-┊│     1#0:o A three
-┊│     1#0:t A two
-┊●   1#1 add one
-┊│     1#1:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_with_duplicate_branch_sources() {
-    let env = two_branches();
-
-    env.but("_squash2 one one -t 1#0 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Squashed branch 'one' to create commit 00e6751
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ se [second]
-┊●   1#0 add four
-┊│     1#0:q A four
-┊│     1#0:k A one
-┊│     1#0:t A two
-┊●   1#1 add three
-┊│     1#1:o A three
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn amend_uncommitted_files_into_commit() {
-    let env = scenario_with_uncommitted_changes();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   kl   A one
-┊   or   A three
-┊   twop A two
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 (no commit message) (no changes)
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("_squash2 one two -t 1 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Amended 7adb8e6 to create d2f176a
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   or A three
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 (no commit message)
-┊│     1:k A one
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-}
-
-#[test]
-fn amend_all_uncommitted_changes_into_commit() {
-    let env = scenario_with_uncommitted_changes();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   kl   A one
-┊   or   A three
-┊   twop A two
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 (no commit message) (no changes)
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("_squash2 zz -t 1 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Amended 7adb8e6 to create 0e76889
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 (no commit message)
-┊│     1:k A one
-┊│     1:o A three
-┊│     1:t A two
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn amend_uncommitted_hunks_into_commits() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    let lines = std::iter::repeat_n("line\n", 10).collect::<Vec<_>>();
-    env.file("file", lines.concat());
-
-    env.but("commit -b my-branch --no-message")
-        .assert()
-        .success();
-
-    env.prepend_file("file", "top");
-    env.append_file("file", "bottom");
-
-    env.but("diff")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-─────────╮
-qs:9 file│
-─────────╯
-     1│+topline
-   1 2│ line
-   2 3│ line
-   3 4│ line
-─────────╮
-qs:d file│
-─────────╯
-    7  8│ line
-    8  9│ line
-    9 10│ line
-   10   │-line
-      11│+bottom
-
-"#]]);
-
-    env.but("_squash2 qs:9 -t 1 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Amended bcf07e2 to create cb08f3a
-
-"#]]);
-
-    env.but("diff")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-─────────╮
-qs:d file│
-─────────╯
-    8  8│ line
-    9  9│ line
-   10 10│ line
-   11   │-line
-      11│+bottom
-
-"#]]);
-}
-
-#[test]
-fn amend_all_uncommitted_changes_when_zz_is_empty() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 zz -t 1#0 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Amended f55169f to create f55169f
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three
-┊│     1#0:o A three
-┊●   1#1 add two
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn amend_committed_file() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0:o -t 1#1 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Amended f63361f to create 5ab5165
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three (no changes)
-┊●   1#1 add two
-┊│     1#1:o A three
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn cannot_amend_files_from_different_commits() {
-    let env = one_branch_three_commits();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three
-┊│     1#0:o A three
-┊●   1#1 add two
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#0:o 1#1:t -t 1#2 -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: All committed files must come from the same commit. Found files from f55169f and f63361f
-
-"#]]);
-}
-
-#[test]
-fn cannot_amend_files_in_ways_that_cause_conflicts() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "file content");
-    env.but("commit -m 'add file'").assert().success();
-
-    env.file("file", "changed");
-    env.but("commit -m 'change file'").assert().success();
-
-    env.remove_file("file");
-    env.but("commit -m 'remove file'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 remove file
-┊│     1#0:q D file
-┊●   1#1 change file
-┊│     1#1:q M file
-┊●   1#2 add file
-┊│     1#2:q A file
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#0:q -t 1#2 -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Failed to apply changes to destination commit - merge conflict
-
-"#]]);
-}
-
-#[test]
-fn squash_into_branch_tip() {
-    let env = one_branch_three_commits();
-
-    env.file("file", "file content");
-
-    env.but("_squash2 file -t a-branch-1 -u")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Amended f55169f to create 13baa98
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three
-┊│     1#0:q A file
-┊│     1#0:o A three
-┊●   1#1 add two
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squash_into_empty_branch() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "file content");
-
-    env.but("branch new bottom").assert().success();
-    env.but("_squash2 file -t bottom -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --target cannot be an empty branch
-
-"#]]);
-
-    // middle and bottom are stil empty even if they're stacked
-    env.but("branch new middle -a bottom").assert().success();
-    env.but("_squash2 file -t middle -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --target cannot be an empty branch
-
-"#]]);
-    env.but("_squash2 file -t bottom -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --target cannot be an empty branch
-
-"#]]);
-
-    env.but("commit --empty -b bottom --no-message")
-        .assert()
-        .success();
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   qs A file
-┊
-┊╭┄ mi [middle] (no commits)
-┊│
-┊├┄ bo [bottom]
-┊●   1 (no commit message) (no changes)
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-    // middle should be considered empty even though there are commits on its parent
-    env.but("_squash2 file -t middle -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --target cannot be an empty branch
-
-"#]]);
-}
-
-#[test]
-fn cannot_squash_into_targets_that_dont_exist() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "file content");
-
-    env.but("_squash2 file -t does-not-exist -u")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Could not find target: 'does-not-exist'
-
-Hint: --target must be an applied commit, branch, or zz. Run `but status` for applicable targets.
-
-"#]]);
-}
-
-#[test]
-fn squash_into_zz_to_uncommit_commit() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0 -t zz")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Uncommitted f55169f
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   or A three
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add two
-┊│     1#0:t A two
-┊●   1#1 add one
-┊│     1#1:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("undo").assert().success();
-
-    env.but("_squash2 1#0 -t zz --format json")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#""#]]);
-}
-
-#[test]
-fn squash_into_zz_to_uncommit_file() {
-    let env = one_branch_three_commits();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three
-┊│     1#0:o A three
-┊●   1#1 add two
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#0:o -t zz")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-Uncommitted from f55169f
-
-"#]]);
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   or A three
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 add three (no changes)
-┊●   1#1 add two
-┊│     1#1:t A two
-┊●   1#2 add one
-┊│     1#2:k A one
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-}
-
-#[test]
-fn cannot_uncommit_files_in_ways_that_cause_conflicts() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "file content");
-    env.but("commit -m 'add file'").assert().success();
-
-    env.file("file", "changed");
-    env.but("commit -m 'change file'").assert().success();
-
-    env.remove_file("file");
-    env.but("commit -m 'remove file'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1#0 remove file
-┊│     1#0:q D file
-┊●   1#1 change file
-┊│     1#1:q M file
-┊●   1#2 add file
-┊│     1#2:q A file
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#2 -t zz")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Cannot uncommit commits that would result in merge conflicts
-
-"#]]);
-
-    env.but("_squash2 1#2:q -t zz")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: Cannot uncommit hunks that would result in merge conflicts
-
-"#]]);
-}
-
-#[test]
-fn cannot_use_source_message_with_uncommitted_changes() {
-    let env = one_branch_three_commits();
-
-    env.file("file", "file content");
-
-    env.but("_squash2 file -t a-branch-1 --use-source-message")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --use-source-message cannot be used when squashing uncommitted changes
-
-"#]]);
-
-    env.but("_squash2 zz -t a-branch-1 --use-source-message")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --use-source-message cannot be used when squashing uncommitted changes
-
-"#]]);
-}
-
-#[test]
-fn cannot_use_source_message_when_moving_committed_files() {
-    let env = one_branch_three_commits();
-
-    env.but("_squash2 1#0:o -t 1#1 --use-source-message")
-        .assert()
-        .failure()
-        .stderr_eq(snapbox::str![[r#"
-Error: --use-source-message cannot be used when moving committed changes
-
-"#]]);
-}
-
-#[test]
-fn committed_file_to_uncommitted_area() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "second commit");
-
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        // .stderr_eq(snapbox::str![""])
-        .stdout_eq(snapbox::str![[r#"
-...
-{
-  "uncommittedChanges": [],
-  "stacks": [
-    {
-      "cliId": "i0",
-      "assignedChanges": [],
-      "branches": [
-        {
-          "cliId": "g0",
-          "name": "A",
-          "commits": [
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#0:n",
-                  "filePath": "a.txt",
-                  "changeType": "modified"
-                },
-                {
-                  "cliId": "1#0:p",
-                  "filePath": "b.txt",
-                  "changeType": "modified"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#1:n",
-                  "filePath": "a.txt",
-                  "changeType": "added"
-                },
-                {
-                  "cliId": "1#1:p",
-                  "filePath": "b.txt",
-                  "changeType": "added"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "t:t",
-                  "filePath": "A",
-                  "changeType": "added"
-                }
-              ]
-            }
-...
-
-"#]]);
-
-    env.but("_squash2 1#0:p -t zz").assert().success();
-
-    // Verify that `status` reflects the move.
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![""])
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "pn",
-      "filePath": "b.txt",
-      "changeType": "modified"
-    }
-  ],
-  "stacks": [
-    {
-      "cliId": "j0",
-      "assignedChanges": [],
-      "branches": [
-        {
-          "cliId": "g0",
-          "name": "A",
-          "commits": [
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#0:n",
-                  "filePath": "a.txt",
-                  "changeType": "modified"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "1#1:n",
-                  "filePath": "a.txt",
-                  "changeType": "added"
-                },
-                {
-                  "cliId": "1#1:p",
-                  "filePath": "b.txt",
-                  "changeType": "added"
-                }
-              ]
-            },
-            {
-...
-              "changes": [
-                {
-                  "cliId": "t:t",
-                  "filePath": "A",
-                  "changeType": "added"
-                }
-...
-    },
-    {
-      "cliId": "k0",
-      "assignedChanges": [],
-      "branches": [
-        {
-          "cliId": "h0",
-          "name": "B",
-          "commits": [
-            {
-...
-              "changes": [
-                {
-                  "cliId": "l:p",
-                  "filePath": "B",
-                  "changeType": "added"
-                }
-              ]
-            }
-...
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_hunk_to_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    // Must set metadata to match the scenario
-    env.setup_metadata(&["A", "B"]);
-
-    commit_file_with_worktree_changes_as_two_hunks(&env, "A", "a.txt");
-
-    let target_cli_id = branch_commit_cli_ids(&status_json(&env)?, "A")[0].clone();
-    // The amended commit is identified by its change ID, from a freshly built
-    // map that knows the post-amend workspace.
-    env.but(format!("_squash2 zz:a.txt:#0 -t {target_cli_id} -u"))
-        .assert()
-        .success();
-
-    // Verify that only one hunk was assigned ("a.txt" still appears in the
-    // uncommitted area because there is one hunk still unassigned).
-    env.but("--format json status -f")
-        .allow_json()
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-{
-  "uncommittedChanges": [
-    {
-      "cliId": "nk",
-      "filePath": "a.txt",
-      "changeType": "modified"
-    }
-  ],
-...
-
-"#]]);
-
-    // Verify that the commit indeed received the hunk.
-    snapbox::assert_data_eq!(
-        env.open_repo()
-            .rev_parse_single("A:a.txt")?
-            .object()?
-            .try_into_blob()?
-            .take_data(),
-        str![[r#"
-firsta
-line
-line
-line
-line
-line
-line
-line
-last
-
-"#]],
-    );
-
-    Ok(())
-}
-
-// Regression: filenames with dashes should not be misinterpreted as ranges.
-// Before the fix, "my-file.txt" would be split on '-' and treated as a range
-// from "my" to "file.txt", which would fail.
-
-#[test]
-fn uncommitted_hunk_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    env.file("uncommitted-to-commit.txt", "content\n");
-
-    let before = status_json(&env)?;
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!(
-        "_squash2 uncommitted-to-commit.txt -t {target_cli_id} -u"
-    ))
-    .assert()
-    .success();
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "uncommitted-to-commit.txt"),
-        "file should no longer be uncommitted"
-    );
-    assert!(
-        branch_commits_contain_file(&after, "A", "uncommitted-to-commit.txt"),
-        "file should appear in commits on branch A"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_area_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    env.file("zz-to-commit.txt", "content\n");
-
-    let before = status_json(&env)?;
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!("_squash2 zz -t {target_cli_id} -u"))
-        .assert()
-        .success();
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "zz-to-commit.txt"),
-        "file should no longer be uncommitted"
-    );
-    assert!(
-        branch_commits_contain_file(&after, "A", "zz-to-commit.txt"),
-        "file should appear in commits on branch A"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_to_commit_consumes_renames() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    let original = (1..=120)
-        .map(|line| line.to_string())
-        .collect::<Vec<_>>()
-        .join("\n")
-        + "\n";
-    env.file("rename-source.txt", &original);
-    env.but("commit -b A -m 'seed rename source'")
-        .assert()
-        .success();
-
-    std::fs::rename(
-        env.projects_root().join("rename-source.txt"),
-        env.projects_root().join("rename-target.txt"),
-    )?;
-    env.file(
-        "rename-target.txt",
-        original.replace("40\n41\n42\n", "40\nchanged\n42\n"),
-    );
-
-    let before = status_json(&env)?;
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!("_squash2 zz -t {target_cli_id} -u"))
-        .assert()
-        .success();
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "rename-target.txt"),
-        "renamed file should no longer be uncommitted"
-    );
-    assert_eq!(
-        env.invoke_git("status --porcelain"),
-        "",
-        "expected all zz changes to be committed"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_file_to_commit_consumes_renames() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    let original = (1..=120)
-        .map(|line| line.to_string())
-        .collect::<Vec<_>>()
-        .join("\n")
-        + "\n";
-    env.file("rename-source-single.txt", &original);
-    env.but("commit -b A -m 'seed rename source single'")
-        .assert()
-        .success();
-
-    std::fs::rename(
-        env.projects_root().join("rename-source-single.txt"),
-        env.projects_root().join("rename-target-single.txt"),
-    )?;
-    env.file(
-        "rename-target-single.txt",
-        original.replace("70\n71\n72\n", "70\nchanged\n72\n"),
-    );
-
-    let before = status_json(&env)?;
-    let source_file_cli_id = uncommitted_cli_id_for_file(&before, "rename-target-single.txt")
-        .expect("renamed uncommitted file should be present in status");
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-
-    env.but(format!(
-        "_squash2 {source_file_cli_id} -t {target_cli_id} -u"
-    ))
-    .assert()
-    .success();
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "rename-target-single.txt"),
-        "renamed file should no longer be uncommitted"
-    );
-
-    let remaining = env.invoke_git("status --porcelain");
-    assert_eq!(
-        remaining, "",
-        "expected selected renamed file to be committed; remaining status:\n{remaining}"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn uncommitted_deleted_file_to_commit_keeps_unrelated_deleted_file() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    env.file("a.txt", "a\n");
-    env.file("b.txt", "b\n");
-    env.file("c.txt", "c\n");
-    env.but("commit -b A -m 'Add a.txt, b.txt, and c.txt'")
-        .assert()
-        .success();
-
-    std::fs::remove_file(env.projects_root().join("a.txt"))?;
-    std::fs::remove_file(env.projects_root().join("b.txt"))?;
-
-    let before = status_json(&env)?;
-    let source_file_cli_id = uncommitted_cli_id_for_file(&before, "a.txt")
-        .expect("a.txt deletion should be present in the uncommitted area");
-    let target_cli_id = branch_commit_cli_ids(&before, "A")[0].clone();
-    assert!(
-        uncommitted_contains_file(&before, "b.txt"),
-        "b.txt deletion should start in the uncommitted area"
-    );
-
-    env.but(format!(
-        "_squash2 {source_file_cli_id} -t {target_cli_id} -u"
-    ))
-    .assert()
-    .success();
-
-    let after = status_json(&env)?;
-    assert!(
-        !uncommitted_contains_file(&after, "a.txt"),
-        "selected a.txt deletion should be amended into the target commit"
-    );
-    assert!(
-        uncommitted_contains_file(&after, "b.txt"),
-        "unrelated b.txt deletion should remain uncommitted"
-    );
-    assert!(
-        !env.projects_root().join("a.txt").exists(),
-        "selected a.txt deletion should stay applied to the worktree"
-    );
-    assert!(
-        !env.projects_root().join("b.txt").exists(),
-        "unrelated b.txt deletion should stay applied to the worktree"
-    );
-    assert!(
-        env.projects_root().join("c.txt").exists(),
-        "untouched c.txt should stay in the worktree"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn commit_to_uncommitted_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    let before = status_json(&env)?;
-    let commit_cli_ids_before = branch_commit_cli_ids(&before, "A");
-    let source_cli_id = commit_cli_ids_before[0].clone();
-
-    env.but(format!("_squash2 {source_cli_id} -t zz"))
-        .assert()
-        .success();
-
-    let after = status_json(&env)?;
-    let commit_cli_ids_after = branch_commit_cli_ids(&after, "A");
-
-    assert_eq!(
-        commit_cli_ids_after.len() + 1,
-        commit_cli_ids_before.len(),
-        "uncommitting a commit should remove that commit from branch history"
-    );
-    assert!(
-        !commit_cli_ids_after.contains(&source_cli_id),
-        "source commit should no longer be present after uncommit"
-    );
-
-    assert!(
-        uncommitted_contains_file(&after, "a.txt") && uncommitted_contains_file(&after, "b.txt"),
-        "uncommitting a commit should move its changes into uncommitted"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn commit_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-    env.setup_metadata(&["A", "B"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "second commit");
-
-    let before = status_json(&env)?;
-    let commit_cli_ids_before = branch_commit_cli_ids(&before, "A");
-    let source_cli_id = commit_cli_ids_before[0].clone();
-    let target_cli_id = commit_cli_ids_before[1].clone();
-
-    env.but(format!("_squash2 {source_cli_id} -t {target_cli_id} -u"))
-        .assert()
-        .success();
-
-    let after = status_json(&env)?;
-    let commit_cli_ids_after = branch_commit_cli_ids(&after, "A");
-    assert_eq!(
-        commit_cli_ids_after.len() + 1,
-        commit_cli_ids_before.len(),
-        "squashing should reduce commit count by one"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn commit_without_message_to_commit() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file("one.txt", "one.txt contents");
-    env.but("commit -m 'add one.txt' one.txt")
-        .assert()
-        .success();
-
-    env.but("status --no-hint")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1 add one.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-"#]]);
-
-    env.but("commit --empty --no-message").assert().success();
-
-    env.but("status --no-hint")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1#0 (no commit message) (no changes)
-┊●   1#1 add one.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-"#]]);
-
-    env.but("_squash2 1#0 -t 1#1 -u").assert().success();
-
-    env.but("status --no-hint")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1 add one.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-"#]]);
-}
-
-#[test]
-fn commit_to_commit_without_message() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file("one.txt", "one.txt contents");
-    env.but("commit -m 'add one.txt' one.txt")
-        .assert()
-        .success();
-    env.but("commit --empty --no-message").assert().success();
-
-    env.but("_squash2 1#1 -t 1#0 --use-source-message")
-        .assert()
-        .success();
-
-    let status = status_json(&env)?;
-    let branch = status["stacks"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .flat_map(|stack| stack["branches"].as_array().unwrap().iter())
-        .find(|branch| branch["name"].as_str().unwrap() == "A")
-        .unwrap();
-    let commit_messages = branch["commits"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|commit| commit["message"].as_str().unwrap().trim_end_matches('\n'))
-        .collect::<Vec<_>>();
-
-    assert_eq!(commit_messages, vec!["add one.txt", "add A"]);
-
-    Ok(())
-}
-
-#[test]
-fn committed_file_to_commit_smoke() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "source-a.txt", "source-b.txt", "source commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "target-a.txt", "target-b.txt", "target commit");
-
-    let before = status_json(&env)?;
-    let source_cli_id = branch_commit_cli_id_for_file(&before, "A", "source-a.txt")
-        .expect("source commit with file");
-    let target_cli_id = branch_commit_cli_id_for_file(&before, "A", "target-a.txt")
-        .expect("target commit with file");
-
-    env.but(format!(
-        "_squash2 {source_cli_id}:source-a.txt -t {target_cli_id} -u"
-    ))
-    .assert()
-    .success();
-
-    let after = status_json(&env)?;
-    let branch = after["stacks"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .flat_map(|stack| stack["branches"].as_array().unwrap().iter())
-        .find(|branch| branch["name"].as_str() == Some("A"))
-        .expect("branch A should remain in status");
-    let commit_contains_file = |message: &str, file_path: &str| {
-        branch["commits"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .find(|commit| {
-                commit["message"]
-                    .as_str()
-                    .is_some_and(|actual| actual.trim_end_matches('\n') == message)
-            })
-            .unwrap_or_else(|| panic!("commit '{message}' should remain in branch A"))["changes"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|change| change["filePath"].as_str() == Some(file_path))
-    };
-    assert!(
-        !commit_contains_file("create source-a.txt and source-b.txt", "source-a.txt"),
-        "moved file should be absent from the rewritten source commit"
-    );
-    assert!(
-        commit_contains_file("create target-a.txt and target-b.txt", "source-a.txt"),
-        "moved file should be present in the rewritten target commit"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn squash_amending_modified_and_renamed_file() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-    env.setup_metadata(&[]);
-
-    env.file("file", "content");
-    env.file("file-2", "content-2");
-
-    env.but("commit -m 'add files'").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-┊│     1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.file("file-2", "new content");
-    env.rename_file("file-2", "file");
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   qs M file
-┊   kw D file-2
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-┊│     1:k A file-2
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("_squash2 zz -t 1 -u").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1]
-┊●   1 add files
-┊│     1:q A file
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn doesnt_open_editor_if_no_sources_have_message() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file(".git/editor.sh", "false");
-    let editor_path = env.projects_root().join(".git/editor.sh");
-    let editor_command = format!("sh {}", editor_path.display());
-
-    env.but("commit --empty --no-message").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1 (no commit message) (no changes)
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1 -t tpm")
-        .env("GIT_EDITOR", editor_command)
-        .assert()
-        .success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn doesnt_open_editor_if_no_target_has_message() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file(".git/editor.sh", "false");
-    let editor_path = env.projects_root().join(".git/editor.sh");
-    let editor_command = format!("sh {}", editor_path.display());
-
-    env.but("commit --empty --no-message").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1 (no commit message) (no changes)
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 tpm -t 1")
-        .env("GIT_EDITOR", editor_command)
-        .assert()
-        .success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1 add A
-┊│     1:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn doesnt_open_editor_if_both_source_and_target_doesnt_have_a_message() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file(".git/editor.sh", "false");
-    let editor_path = env.projects_root().join(".git/editor.sh");
-    let editor_command = format!("sh {}", editor_path.display());
-
-    env.but("commit --empty --no-message").assert().success();
-    env.but("commit --empty --no-message").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1#0 (no commit message) (no changes)
-┊●   1#1 (no commit message) (no changes)
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    env.but("_squash2 1#0 -t 1#1")
-        .env("GIT_EDITOR", editor_command)
-        .assert()
-        .success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1 (no commit message) (no changes)
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
-
-#[test]
-fn squashing_into_branch_that_sits_below_empty_branch() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-    env.setup_metadata(&["A"]);
-
-    env.file("file", "content");
-
-    env.but("branch new -a A").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted]
-┊   qs A file
-┊
-┊╭┄ br [a-branch-1] (no commits)
-┊│
-┊├┄ g0 [A]
-┊●   tpm add A
-┊│     tpm:t A A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
-
-"#]]);
-
-    env.but("_squash2 file -t A").assert().success();
-
-    env.but("status -f")
-        .assert()
-        .success()
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ br [a-branch-1] (no commits)
-┊│
-┊├┄ g0 [A]
-┊●   tpm add A
-┊│     tpm:t A A
-┊│     tpm:q A file
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-}
